Add script to extract content from the radiant install
This spits out a bunch of <slug>/index.html.<filter_extension> files nested according to page structure from radiant.  Currently it spits out each non-body part as a partial, but that might not be the right thing to do.

It is expected to be run by script/runner from the main radiant install (e.g. what's currently on master).