Extract a meeting_pages helper
Instead of using the query interface we just use sitemap.resources and the standard ruby enumerator API for filtering.  In part because we needed to filter in places on things not present in the data for the page, and in part because it appears that middleman 4 will drop the query interface anyway.