Force success on the pre-roll build step
We build the app before we deploy (we should review this as it might be overkill, but we took most of this deploy.rb from elsewhere).  We currently have errors in the build and thus the middleman command gives a non-zero exit status and this makes ssh_kit break.  We don't actually care so we'll use shell script tricks "<cmd>; true" to force success - this might not be the cleanest approach.