Move build/* back into <release_path>/ and remove <release_path>/build
We tar from the app root, but only the build dir, this ends up being deployed to the release path and preserves the build dir.  We don't want that so we move everything out of build and back down one folder and then remove the build folder.

We also have to deal with the ._<blah> resource fork things that taring files on OS X will cause.  We remove these first before moving stuff and then do rmdir.  We could avoid this and use rm -rf, but our way means if there is some other dot-file that's added but not moved directly we'll fail the deploy rather than delete it silently.