Move members-and-friends over to middleman
We put the member list into data/ so it's easy for folk to add themselves with a PR if they want.  We also duplicate the /as.xml version of the page, although I have no idea if anyone is actually using it, so maybe we don't need to keep it around?

Part of #12