Hide files that don't have status Published in their frontmatter
We use a ready hook in the config.rb to scan through the sitemap and explicitly ignore any files without frontmatter that sets them as published.  We also explicitly keep .htaccess files and anything taht looks like an asset (for now that's anything in javascripts/, stylesheets/, or images/).

I guess this is laying some groundwork to make #18 and #19 more worthwhile.