Specify redirects as data and add them as RewriteRules to .htaccess
We had to make .htaccess be ERB evaluated to do this and this meant it came out as .htaccess.html, so we had to use a proxy rule in config.rb to keep it around.  Apparently this'll go away in middleman v4, which will be nice.

Our initial data is the 4 redirects from radiant, so this closes #13