Remove old tips-for-speakers page and replace with a redirect
As noted in redirects.json we'll redirect to /speaking/tips which will in turn be redirected to readme.lrug.org/#talks - it might be better to redirect directly, but you never know - we might want /speaking/tips to go elsewhere later so this gives us that flexibility.

Part of #13 and #12 I guess.