Remove final pages from lrug_root
We've already dealt with meetings and rss-feeds in previous commits.

The sekrit-notes were unpublished pages that I used as a notepad for organising meetings or giveaways.  These don't need to be published in the new site, nor do they need to live in archive in case we ever need the content again.