Replace complicated rss link in layout with rss_path helper
As before all pages will have a link to /rss/ unless conditions are met, but instead of looking at the url we instead look at the ``content_for :rss_path`` which we set in the layouts or index pages as appropriate.  It does spread the knowledge out a bit to the pages and layouts, but it makes the base layout cleaner.  It's a bit cumbersome that we can set :rss_path in the meetings layout for all the meeting pages, but have to set it manually in the /meetings/index and /meetings/<year>/index pages (similar for podcasts).