Allow for data-driven rendering of sponsors on meeting pages
We still allow a fully custom data.part.sponsors to be rendered as a chunk of HTML but if there is a data.sponsors we will render a simple list from that.  This means most sponsors are simple enough, but for when we have multiple we can still override.

In the cases where we do override, we still add data.sponsors because it'll come in handy later for making /sponsors more dynamic.