Provide sponsor extension for kramdown
We can write `{::sponsor name="foo" size="bar" /}` in a markdown document and it will render out the bar sized logo for the foo sponsor from data/sponsors.json.  We provide a new kramdown parser that supports this extension (lib/lrug_extended_kramdown.rb) and this gives us scope for other extensions in the future.

Upgrade all the places that were using ERB or rendering the sponsor logos manually to use this new construct.