Add description metadata for Twitter and Facebook
As we don't have any description / summary frontmatter (yet?) I took some inspiration from how middleman-blog does it[1].  We render the page and extract the first `p` tag.  This'll do for now, but it does mean that the "unfurl" in slack for meeting pages usually ends with "Full venue and registration details are given below." and there's no "below" to speak of.

It would be nice if we could get meeting pages to provide custom descriptions to avoid this.  The article I was following[2] suggests we can use twitter card label and data properties to add custom things to the unfurl in slack so I thought we could add some frontmatter and use that to include registration links using these properties.  However the twitter card documentation for those tags[3] says they were deprecated over a year ago, so it seems a non-starter.

[1]: https://github.com/middleman/middleman-blog/blob/1f154d8846e22bfa846db0b13ada1f8c90f6d213/lib/middleman-blog/truncate_html.rb#L10-L14
[2]: https://medium.com/slack-developer-blog/everything-you-ever-wanted-to-know-about-unfurling-but-were-afraid-to-ask-or-how-to-make-your-e64b4bb9254#.tu7nlzj2m
[3]: https://dev.twitter.com/cards/types/product