Refer to data in config.rb via `@app`
We need to tell our kramdown extension about the list of sponsors that
we've defined as data so it can render their logos.  Unfortunately if we
refer to `data.sponsors` this doesn't get us the data, but a
Middleman::CoreExtensions::Collection::LazyCollectorStep which is some
kind of late binding object that does contain the info, but not directly.

I had a look at the following issue
https://github.com/middleman/middleman/issues/1810 to get some tips.  The
first suggestion, referring to `data` in an `after_configuration` block,
didn't seem to work.  Happily the second one, refer to `data` directly on
the `@app` object instead of the global version, did work.  It's unclear
from the issue, or any other documentation if this is expected, or if the
lazy thing is a bug.