Stop invoking Tilt directly for non-page renders
We have a helper that renders markdown, and a helper that can render a
"part" from the frontmatter.  For the "part" renderer we also allow
setting multiple filters on the part and our helper will pass the content
through a renderer for each filter type, much like middleman itself does
with multiple file extensions on templates (e.g. middleman will pass
a file called foo.html.md.erb to the ERB renderer and then the Markdown
renderer; setting the filter on a part to ".md.erb" will do the same).

The problem is that we were invoking Tilt directly to do this and in
middleman 4 this meant it was no longer picking up the context of the
middleman app (so had no access to helpers) and no longer picked up
the template engine config settings from config.rb.  Fixing the former
is simple, we can provide `context: @app` to Tile for rendering, but
providing the config settings is harder before the `options_for_ext`
method we used to use is no longer public.  Instead we look at what
Middleman::TemplateRenderer does and crib off it: create a
Middleman::FileRenderer and give it the application context.  This
looks at the last extension on the file name to work out the renderer
to use, and can be given the body to parse, rather than reading it
from disk.  We create a fake filename with the extensions of all the
renderers we want (from the filter config of the part) and pass that
to the file renderer.  We iterate over each filter removing the last
extension, and passing the body content explicitly until we're done.

This is very similar to how middleman does it in the TemplateRenderer
except it doesn't provide the template body for the first iteration so
the FileRenderer reads it from disk, whereas we provide the template
body on the first iteration from the value of the part content in the
frontmatter.

It does seem like this is a bit painful, and could be something that
middleman provides out of the box.