Rely on default travis/dpl
We were relying on lazyatom's fork of dpl so we could get the "overwrite" option
to let us upload our build artefact to the same file on the same branch each
time.  However, that [has been merged to upstream][travis-dpl-merge] so we no
longer need to rely on the branch.

[travis-dpl-merge]: https://github.com/travis-ci/dpl/pull/386