Add data of coverage links extracted from lanyrd
When we used lanyrd for the meetings we would add coverage links to the talks.
This info basically ones't exist anywhere else and unfortunately lanyrd has
slowly rotted off the web.  I wrote a scraper the last time it appeared online
to get the coverage links and this data file is the result.  It's incomplete
as some pages are missing from the scrape.  It's also incomplete as we only
added meetings to lanyrd regularly from 2011-2016 and even then we didn't add
all the coverage.

Still, it's a good start.