Add a `{::coverage}` kramdown extension
This helper lets us render data from the coverage.json file as a series of
links.  The arguments are the meeting (in form year/month-name) and the talk
itself (in the form of a parameterised id; downcase and strip non alphanumeric
chars).  It's a bit annoying that we have to pass the meeting in, rather than
extracting it implicitly from the context of the page being rendered.  The
problem here is that the kramdown renderer doesn't really have that context.