Add example coverage links for April and May 2011 talks
If this looks ok we'll (manually, ugh) add what we have for all the other
meetings and talks that we have coverage for.  Once that's complete we'll then
want to scour the mailing list and twitter for any links we might have to
coverage for meetings before we used lanyrd and after we stopped using lanyrd.