Stop relying on Skills Matter as a default host
1. Make `hosted_by` a bit of first-class data in the frontmatter instead
   of being a content part
2. Remove the default `hosted_by` from the `meetings/index`
3. Change the sponsors snippet to prefer a non-inherited `hosted_by`
   content part, then a list of `hosted_by` sponsors as data.  It raises
   if the meeting doesn't define a host.
4. Update all the existing meetings to have correct `hosted_by` data.

Like `sponsors` we've left in the option of having a `hosted_by` content
part if for some reason the hosted by information shouldn't just be a list
of named sponsors, although we've not needed it so far.

For the Jan 2007 meeting where we self-hosted a quiz in a pub I've put
LRUG itself in as the host, and added an entry for LRUG to the
`sponsors.json` data file.  Feels a bit weird, but whatever.