Update sponsors page to reflect new `hosted_by` structure
Now that `sponsors` and `hosted_by` data is the same shape the two methods
can do the same thing.  I've tweaked it to sort by occurences so that the
companies that have sponsored the most appear at the top of the list.

For now I've also filtered LRUG and Skills Matter from the list of hosts
as we've already called out SM manually above (and I can imagine we'd
continue to do so) and it doesn't make sense to put ourselves in that
list.