Restructure `_sponsors.erb` to allow for no host
Now we've made sure all the old meetings that were implicitly hosted by
Skills Matter are explicitly marked as such, we don't need to raise an
exception if a meeting has no host.  Turns out, this will be useful in
exactly 1 commit's time because the may 2020 meeting won't have a host
because we're running on a Zoom account we've paid for.