Replace `filter` with `render_as` in frontmatter
While running the build to check the upgrades in the previous commit
hadn't broken anything, I noticed the following warning message:

    WARN -- : You are setting a key that conflicts with a built-in method Middleman::Util::EnhancedHash#filter defined in Hash. This can cause unexpected behavior when accessing the key as a property. You can still access the key via the #[] method.

and finally decided to do something about it.  The hint is the reference
to `Middleman::Util::EnhancedHash#filter` and by searching for `filter` I
realised we used it in frontmatter as a hint for how to render chunks of
content defined in the frontmatter.  For example, some sidebar content
defined in the main index pages is written in markdown so we want to let
the system know that.  The name `filter` isn't hugely important so we can
replace it with `render_as` which is proably even clearer and doesn't
collide with `Middleman::Util::EnhancedHash`.