Remove support for `extended` content parts
We don't use them so there's no point.  The radient version used extended
parts to support summaries in the index and the full content on individual
pages, but we never used them.  Or we collapsed them down as part of the
migration.  Either way, we don't use them so lets remove support for them.