Use `main` and `aside` in base layout
We've a `div#main` that might as well become `main#main` and the
`div#sidebar` is clearly an `aside#sidebar`.  Considered converting
`div#content` into a `section#content` or `article#content` but it really
doesn't seem like it makes sense, particularly as we'd sometimes render
lots of `article`s inside it (e.g. on index pages), so I left it as a
`div`.