Get rid of `vevent` divs
Some previous version of the site had manual markup to support hcard
microformats on the page for each meeting.  They're long gone, just
leaving the vestigial tail of a `div.vevent` wrapping the content of each
meeting or podcast on the individual pages.  Microformats were neat, but
without full support in an editor we're never going to mark them up in a
markdown doc.