Add all existing coverage to pages
I wrote a script to generate all the `{::coverage ... /}` tags from the
non-empty entries in the JSON file, and the append those links to the
relevant source file for the meeting.  Then I went through each edited
file and moved the tags to the right place in the source.

Had to tidy up the meetings with lots of speakers that used to just use a
list for formatting too.