Add anchor links to all the kramdown rendered headers
We hide them on the sidebar though because they look weird.

We get this by changing the HTML Kramdown outputs for heading elements to
add a `<a href="#id">#</a>` element inside the header.

Unlike our `LRUGExtendedKramdown` parser, we can't configure a custom
renderer so we have to open `Middleman::Renderers::MiddlemanKramdownHTML`
and monkey-patch it directly.  Bit annoying.