Split coverage into multiple (yaml) files
The JSON coverage file is so long that it's becoming hard to edit. I
think this is partly because it's one big file, and partly because
cleanly-formatted JSON is quite verbose.

This splits the file into one per year, and changes the format to YAML,
which is much terser, but also in this case easier to work with.