Put a redirect in place for styles.css
For better or worse readme.lrug.org takes styles from lrug.org and expects them to live at styles.css.  They haven't for a while, _and_ we recently put in place asset hashes so it's not a static URL.  However, we can use some redirection to make it so that it is.