Take advantage of travis's default bundle install for caching
We specified our own bundle install step which means we don't install to the place that travis expects and so the cache step doesn't give us any advantage what-so-ever as we're not caching where our gems are.  Removing our install step (and thus using the default travis one for ruby + bundler) means we do install to the same place we're caching from and so our builds _should_ get faster.

See https://docs.travis-ci.com/user/caching/#bundler for more detail.