Upgraded middleman to 4.4.2
Mostly to get rid of the dep warnings about `URI.escape` which we get as a hint about ruby 3 compatibility from ruby 2.7:

    /path/to/lrug.org/.bundle/gems/ruby/2.7.0/gems/middleman-core-4.3.11/lib/middleman-core/extensions/asset_hash.rb:90: warning: URI.escape is obsolete

Against my better judgement I just ran `bundle update middleman` rather than `bundle update middleman --conservative` so this upgraded as many gems as possible.  The only issue seems to have been that Haml 6 isn't compatible with how middleman integrated Haml 5 as we get:

    NameError: uninitialized constant Haml::Options

from:

    /path/to/lrug.org/.bundle/gems/ruby/2.7.0/gems/middleman-core-4.4.2/lib/middleman-core/renderers/haml.rb:49:in `initialize'

We don't even use haml so it's not a problem if we're not on the latest and greatest, so I just removed that upgrade.  Doing this got it to run and it generates the site fine in that running `bundle exec middleman build` on this branch vs. the master branch reported that all files were identical (other than `version.json` which we'd expect), so I assume all good.