Remove references to “our own little bubbles” for in-person meetings
This made sense for virtual meetings but is a confusing non sequitur now
that we’ve returned to in-person.

Arguably we should also return to saying more about the scope of the
code of conduct with respect to the socialising part of the evening, per
e.g. the last pre-pandemic announcement [0], but that’s a separate
decision.

[0] https://lrug.org/meetings/2020/march/#afterwards