Use github actions to build + deploy
Mostly this is taking the things we did in the travis deploy and doing
them via github actions using [github script][] to use the provided
octokit/rest instance to:

* delete the lrug.org.tar.bz2 asset on the travis-release release - if
  it's there already
* upload the new lrug.org.tar.bz2 file from the previous (bash) step to
  the travis-release script

We do this rather than using existing actions because the github actions
we'd use are deprecated and we don't need all the bells and whistles of
the alternatives. We upload the same named asset to the same release each
time.

Of minor interest is that instead of using
`github.rest.repos.uploadReleaseAsset` we use a raw `github.request` to
POST because that means we can set the headers appropriately for the
content type of the data. Unsure if this is necessary TBH. The docs for
[uploading an asset][] say you should, but [the docs for octokit/rest][]
which implement of this API doesn't provide any hint as to how (although
it does duplicate the same info about setting headers as the raw API docs
🤷).

[github script]: https://github.com/actions/github-script
[uploading an asset]: https://docs.github.com/en/rest/releases/assets?apiVersion=2022-11-28#upload-a-release-asset
[the docs for octokit/rest]: https://octokit.github.io/rest.js/v19#repos-upload-release-asset