Add `meeting_date` to frontmatter of meetings
Most generated via a script to extract the date from the opening paragraph as we're mostly fairly consistent with how we write that.  A few where we were more flamboyant I did manually and that means there's some extra whitespace tidying in the commit because my editor does that.  Sorry.