Get back to middleman 4
For some reason dependabot decided to downgrade us from 4 to 0.14 as part of security patching nokogiri from 1.14 to 1.15 - this corrects that.  We also pull back from ActiveSupport and Haml upgrades that we weren't expecting.  We're getting a bunch of other bumps that we maybe don't expect, but the build seems to still run so it's probably fine.