Move lrug_helpers from lib/ to helpers/
This is how we're supposed to create helpers in middleman, and it means we no
longer need to explicitly require it in config.rb which is nice.

This does mean renaming the helper from `LRUGHelpers` to `LrugHelpers` as the
default loading expects a standard snake_case -> CamelCase pattern for finding
module names when loading files and `lrug_helpers` doesn't become `LRUGHelpers`
using this pattern.