Extract talks to data and render via ERB
Take the talks content from the February 2025 markdown file and extract them to
a yaml data file.  We also move the coverage links for those talks from
`data/coverage` to `data/talks` so that all the data for a talk is in one
place.

Because we want to render markdown content, which seems hard from inside a
markdown extention we also change from using a custom kramdown extension tag to
render coverage links to just using ERB. This does mean changing the february
meeting from a `.html.md` extension to a `.html.md.erb` extension so it can be
processed in multiple stages.

The main thing we want to achieve here is being able to export the yaml file for all our talks that can be consumed by the data ingest of rubyevents.  Doing so will happen in a separate commit.