Pass through template options when rendering talks partial
This way on indexes we get the auto_id_prefix that is set in the
meetings/meeting template the talks are rendered on. Meaning on index pages the
talk headings have the meeting id as a prefix just in case we need to
disambiguate two talks with the same title for different meetings. Like when
one is cancelled and moved to the following meeting.

It doesn't feel like this is what we should have to do (reach into the render
object and dup some options from an instance variable), but at the same time I
think erb helpers that invoke partial renders is probably unusual so we'll make
do.