Escape any html present in coverage titles
Thanks to @fcheung for pointing this out. We want to avoid a title with < or >
breaking the markup. Turns out this is still a problem in the markdown
extension version, but we're gonna deprecate that version eventually so let's
just leave it, rather than working out how to get the `h` helper into that same
context.