Use `published_at` vs `announced_at` correctly
`published_at` is the time the _video_ was published whereas `announced_at` is
when the talk was announced. We don't really track publication of the coverage
links (although we probably _could_ extract it from git timestamps on the
coverage yml files). For now populate `announced_at` with the publication date
of the meeting page, and `published_at` with a "TODO" placeholder for videos or
"Not published" for talks we definitely don't have videos for.