Add placeholder for Dec 2025 meeting (#423)
* Add placeholder for Dec 2025 meeting

We're _pretty_ sure it'll be hosted by intercom so copied the details over from
Dec '24 when we were there last. HAven't done the ticket-tailor page yet so
will need to revisit when we have talks + that set up.

* Add the Ticket Tailor link for Dec 2025

Turns out, if you visit a venue once a year on the same month, it's _pretty_
easy to create the ticket tailor page for it.