Fix showing no talks message (#424)
Problem:
`if talks` just checks the talks variable has a value. An empty array
counts as having a value, so we do not render the no talks message.

Fix:
By doing `if talks.any?`, we check if there are talks, and if there are
none, we do show the intended no talks message.