Support missing as well as empty talks
In #424 we replaced a nil check with an `any?` check. This was the right thing
to do given we'd added an empty list of talks for December 2025 to
`talks/2025.yml`, but if we did just leave it out completely we wouldn't get an
empty `[]` we'd get `nil` so we need both checks.