Support missing as well as empty talks
In #424 we replaced a nil check with an `any?` check. This was the right thing
to do given we'd added an empty list of talks for December 2025 to
`talks/2025.yml`, but if we did just leave it out completely we wouldn't get an
empty `[]` we'd get `nil` so we need both checks.  Which we can achieve with
activesupport's `present?` and get some extra "but the array is also full of
empty things" checks too, just in case.

Co-authored-by: Alessandro Proserpio <proserpioalex@gmail.com>